Migrated the main objects representing a Snapshot from org.json to dsl-json to improve performance, new object introduced for this type of deserialization

+++ src/main/java/hu/tibipi/bumbitrack/core/Place.java
@@ -1,14 +1,14 @@
 package hu.tibipi.bumbitrack.core;
 
 public class Place {
-    private final float longitude;
-    private final float latitude;
-    Place(float lo, float la){
+    private final double longitude;
+    private final double latitude;
+    Place(double lo, double la){
         longitude = lo;
         latitude = la;
     }
 
-    double degreesToRadian(float deg){
+    double degreesToRadian(double deg){
         return deg * 3.14159 / 180;
     }
 
@@ -16,7 +16,7 @@ public class Place {
         double latdiff = degreesToRadian(p1.latitude - p2.latitude);
         double lngdiff = degreesToRadian(p1.longitude - p2.longitude);
 
-        float radius = 6371;        //estimation of earth's radius
+        double radius = 6371;        //estimation of earth's radius
 
         double a = Math.sin(latdiff / 2);    //shorthands
         double b = Math.sin(lngdiff / 2);
